+++ docs/components/demos/DateTimePicker.jsx
@@ -1,8 +1,8 @@
 'use strict';
 var React = require('react')
   , dates = require('date-arithmetic')
-  , Button = require('../../bootstrap').Button
-  , ButtonGroup = require('../../bootstrap').ButtonGroup
+var Button      = require('react-bootstrap/lib/Button')
+var ButtonGroup = require('react-bootstrap/lib/ButtonGroup')
   , RW = require('../../../src/index');
 
 require('globalize/lib/cultures/globalize.culture.en-GB');
@@ -62,7 +62,7 @@ module.exports = React.createClass({
                 <div className='checkbox'>
                   <label>
                     <input type='checkbox'
-                      checked={this.state.isRtl}
+                      checked={!!this.state.isRtl}
                       onChange={this._set.bind(null, 'isRtl', !this.state.isRtl)}/>
                       Right to Left
                   </label>