+++ docs/components/demos/MultiSelect.jsx
@@ -1,7 +1,7 @@
 'use strict';
 var React = require('react')
-  , Button = require('../../bootstrap').Button
-  , ButtonGroup = require('../../bootstrap').ButtonGroup
+var Button      = require('react-bootstrap/lib/Button')
+var ButtonGroup = require('react-bootstrap/lib/ButtonGroup')
   , RW = require('../../../src/index')
   , genData = require('../generate-data');
 
@@ -20,8 +20,7 @@ module.exports = React.createClass({
   },
 
   render: function(){
-    var allVals = this.state.value
-      , disabled = this.state.disabled === true || Array.isArray(this.state.disabled);
+    var disabled = this.state.disabled === true || Array.isArray(this.state.disabled);
 
     var props = {
       data: list,
@@ -70,7 +69,7 @@ module.exports = React.createClass({
             <div className='form-group'>
               <label className='checkbox-inline'>
                 <input type='checkbox'
-                  checked={this.state.isRtl}
+                  checked={!!this.state.isRtl}
                   onChange={this._set.bind(null, 'isRtl', !this.state.isRtl)}/>
                   Right to Left
               </label>
@@ -78,13 +77,13 @@ module.exports = React.createClass({
             <div className='form-group'>
               <label className='checkbox-inline'>
                 <input type='checkbox'
-                  checked={this.state.allowCustom}
+                  checked={!!this.state.allowCustom}
                   onChange={this._set.bind(null, 'allowCustom', !this.state.allowCustom)}/>
                   Allow custom tags
               </label>
               <label className='checkbox-inline'>
                 <input type='checkbox'
-                  checked={this.state.groupBy}
+                  checked={!!this.state.groupBy}
                   onChange={this._set.bind(null, 'groupBy', !this.state.groupBy ? 'last' : null )}/>
                   Group
               </label>