+++ docs/components/docs.jsx
@@ -6,10 +6,12 @@ var {
   , Router
   , Navigation
   , State
+  , hashHistory
   , Link } = require('react-router')
 
-var Affix          = require('../bootstrap').Affix
-var Navbar         = require('./topnavbar.jsx')
+var Affix = require('react-overlays/lib/AutoAffix')
+var Navbar = require('./Navbar')
+var Nav = require('react-bootstrap/lib/Nav')
 var GettingStarted = require('./pages/GettingStarted.md')
 var DropdownList   = require('./pages/DropdownList.api.md')
 var ComboBox       = require('./pages/Combobox.api.md')
@@ -24,13 +26,11 @@ var Locale         = require('./pages/i18n.md');
 var Controllables  = require('./pages/controllables.md');
 var OtherProjects  = require('./pages/OtherProjects.md');
 
-var history = require('react-router/lib/HashHistory').history
 
+require('@monastic.panic/component-playground/codemirror.css')
+require('@monastic.panic/component-playground/themes/oceanic.css')
 
-require('../vendor/codemirror.css')
 require('../vendor/styles.css')
-require('../vendor/oceanic-prism.css')
-require('../vendor/oceanic-codemirror.css')
 require('react-widgets/less/react-widgets.less')
 require('../docs.css')
 
@@ -62,35 +62,33 @@ var Docs = React.createClass({
       this.setState({ sideHref: location.hash.split('/')[0] })
   },
 
-  render: function(){
-    var href = this.state.sideHref;
-
+  render() {
     return (
       <div>
         <Navbar page={this.props.page}/>
         <div className='container'>
           <aside className='col-sm-3 section'>
-            <Affix className='nav-aside section-inner' offsetTop={52}>
-              <nav className='side-nav'>
-                <ul className='nav'>
+            <Affix viewportOffsetTop={20}>
+              <div className='nav-aside section-inner'>
+                <nav className='side-nav'>
+                  <Nav>
+                    <li className={'active'}>
+                      <Link to='/getting-started'>Getting Started</Link>
+                    </li>
+                    <li><Link to='i18n'>Localization</Link></li>
+                    <li className='side-divider'>API</li>
+                    <li><Link to='/dropdownlist'>Dropdown List</Link></li>
+                    <li><Link to='/combobox' href='#combobox'>Combobox</Link></li>
+                    <li><Link to='/numberpicker' href='#number-picker'>Number Picker</Link></li>
+                    <li><Link to='/multiselect' href='#multiselect'>Multiselect</Link></li>
+                    <li><Link to='/selectlist'>Select List</Link></li>
+                    <li><Link to='/calendar'>Calendar</Link></li>
+                    <li><Link to='/datetime-picker'>{'Date & Time Picker'}</Link></li>
 
-                  <li className={'active'}>
-                    <Link to='/getting-started'>Getting Started</Link>
-                  </li>
-                  <li><Link to='i18n'>Localization</Link></li>
-                  <li className='side-divider'>API</li>
-                  <li><Link to='/dropdownlist'>Dropdown List</Link></li>
-                  <li><Link to='/combobox' href='#combobox'>Combobox</Link></li>
-                  <li><Link to='/numberpicker' href='#number-picker'>Number Picker</Link></li>
-                  <li><Link to='/multiselect' href='#multiselect'>Multiselect</Link></li>
-                  <li><Link to='/selectlist'>Select List</Link></li>
-                  <li><Link to='/calendar'>Calendar</Link></li>
-                  <li><Link to='/datetime-picker'>{'Date & Time Picker'}</Link></li>
-
-                  <li><Link to='/others'>{'Other projects'}</Link></li>
-                  {/*<li><Link to='advanced'>Advanced</Link></li> */}
-                </ul>
-              </nav>
+                    {/*<li><Link to='advanced'>Advanced</Link></li> */}
+                  </Nav>
+                </nav>
+              </div>
             </Affix>
           </aside>
           <article className='col-sm-9 section'>
@@ -102,14 +100,14 @@ var Docs = React.createClass({
       </div>
     )
   },
-  prev: function(){
+  prev() {
     var idx = locations.indexOf(this.state.sideHref)
       , href = locations[Math.max(idx - 1, 0)];
 
     this.navigate(href)
   },
 
-  next: function(){
+  next() {
     var idx = locations.indexOf(this.state.sideHref)
       , href = locations[Math.min(idx + 1, locations.length - 1)]
 
@@ -118,29 +116,12 @@ var Docs = React.createClass({
 
   handleNavItemSelect: function (key) {
     this.transitionTo(key)
-  },
-
-  navigate: function(){
-    // var change = this.state.sideHref.split('/')[0] !== href.split('/')[0]
-    // this.setState({ sideHref: href });
-    // window.location = href;
-    // if(change)
-    //   window.scrollTo(0, 0)
   }
 })
 
 
-export class Thing extends React.Component {
-  render() {
-    return (
-      <div></div>
-    );
-  }
-}
-
-
 ReactDOM.render((
-  <Router history={history}>
+  <Router history={hashHistory}>
     <Route path="/" component={Docs} indexRoute={{ component: GettingStarted }}>
       <Route path='getting-started(/:topic)' component={GettingStarted}/>